requirements.txt
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
gspread>=5.12.0
google-auth>=2.22.0